Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -4,7 +4,7 @@ int main() {
     float y;
     scanf("%f", &x);
     scanf("%f", &y);
-    int Product= x * y;
+    float Product= x * y;
     printf("Product: %f",x, y,Product);
     return 0;
 }
